requirements.txt
asyncio
python-dotenv
pymongo
requests
discord.py
typing
flask